refactor: split tui into modules
