+++ src/tui/mod.rs
@@ -1,0 +1,5 @@
+pub mod app;
+mod handlers;
+mod render;
+
+pub use handlers::run_tui;
